minor fixes

+++ module/stg.h
@@ -19,5 +19,5 @@
 uint openBmps(char **filePaths, struct BmpStorage *bmpS);
 void closeBmps(struct BmpStorage *bmpS);
 
-int bsEncode(const uint8 *data, uint size, uint position, struct BmpStorage *bmpS);
-int bsDecode(uint8 *data, uint size, uint position, struct BmpStorage *bmpS);
+int bsEncode(const uint8 *data, uint size, loff_t position, struct BmpStorage *bmpS);
+int bsDecode(uint8 *data, uint size, loff_t position, struct BmpStorage *bmpS);
